requirements.txt
# Loads env variables
python-dotenv

# PDF reading
PyMuPDF

# Vector DB
faiss-cpu

# Gemini (Google AI)
google-generativeai


